requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.1.2
    # via streamlit
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.7.0
    # via streamlit
cachetools==5.3.2
    # via streamlit
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via streamlit
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.19.2
    # via altair
jsonschema-specifications==2023.7.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.1
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==23.2
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==2.1.3
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==10.1.0
    # via streamlit
plotly==5.18.0
    # via -r requirements.in
protobuf==4.25.0
    # via streamlit
pyarrow==14.0.1
    # via streamlit
pydeck==0.8.1b0
    # via streamlit
pygments==2.16.1
    # via rich
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
pytz==2023.3.post1
    # via pandas
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via streamlit
rich==13.6.0
    # via streamlit
rpds-py==0.12.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.28.1
    # via -r requirements.in
tenacity==8.2.3
    # via
    #   plotly
    #   streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.3
    # via streamlit
typing-extensions==4.8.0
    # via streamlit
tzdata==2023.3
    # via pandas
tzlocal==5.2
    # via streamlit
urllib3==2.0.7
    # via requests
validators==0.22.0
    # via streamlit
zipp==3.17.0
    # via importlib-metadata
